Add map, a basic layer and legend.
- Add a map with react-map-gl and deck.gl, for now the appearance
is completely temporary (i.e. it shouldn't cover a whole wall) for tests
- Added some basic processing of luftdata to a format that is easier
to work with for the deck.gl layer(s).
- Add a basic layer to show what the circles color mean. Still a work
in progress, needs more detailed labels (good, bad, etc) and to be
even more generalized so that it can be swapped out when switching
layers. Should potentially be able to show multiple layers legend.